% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{This is my addition function}
\usage{
add(x, y)
}
\arguments{
\item{x}{this is the first value to add}

\item{y}{this is the second value to add}
}
\value{
This function returns the sum of x and y
}
\description{
This is my addition function
}
\examples{
## Start with something simple
add(1,1)

## Now something more difficult
add(49,60)

}
