% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.package.R
\docType{package}
\name{my.package}
\alias{my.package-package}
\alias{my.package}
\title{my.package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
