% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_type_destintation_countries.R
\name{pop_type_destintation_countries}
\alias{pop_type_destintation_countries}
\title{Population Type by Destination Countries}
\usage{
pop_type_destintation_countries(
  country1 = NULL,
  country2 = NULL,
  country3 = NULL,
  country4 = NULL,
  country5 = NULL
)
}
\arguments{
\item{country1}{Country Selection}
}
\value{
Graph
}
\description{
Population Type by Destination Countries
}
\examples{
# Example with your data set in "inst/"
datafile_2 <- system.file("unhcr-situations.csv", package = "my.package")
unhcr_situations <- read.csv(datafile_2, encoding = "UTF-8")

# Apply my function
pop_type_destintation_countries('Angola', 'zimbabwe', 'zambia')
}
