% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_function.R
\name{square_function}
\alias{square_function}
\title{square_function}
\usage{
square_function(x)
}
\arguments{
\item{x}{any real number}
}
\value{
the square of x
}
\description{
square_function
}
\examples{
square_function(2)
}
