% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{This is my subtract function}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{this is the first value}

\item{y}{this is the second value to subtract}
}
\value{
This function returns the difference of x and y
}
\description{
This is my subtract function
}
\examples{
## Start with something simple
subtract(1,1)

## Now something more difficult
subtract(49,60)

}
